# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CountByLabel": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "count"
        ],
        "title": "CountByLabel",
        "type": "object"
      },
      "DashboardSummary": {
        "properties": {
          "closed_this_month_count": {
            "title": "Closed This Month Count",
            "type": "integer"
          },
          "due_this_week_count": {
            "title": "Due This Week Count",
            "type": "integer"
          },
          "in_progress_count": {
            "title": "In Progress Count",
            "type": "integer"
          },
          "open_count": {
            "title": "Open Count",
            "type": "integer"
          },
          "overdue_count": {
            "title": "Overdue Count",
            "type": "integer"
          },
          "pending_count": {
            "title": "Pending Count",
            "type": "integer"
          },
          "reopened_count": {
            "title": "Reopened Count",
            "type": "integer"
          },
          "stagnant_count": {
            "title": "Stagnant Count",
            "type": "integer"
          },
          "total_active_count": {
            "title": "Total Active Count",
            "type": "integer"
          }
        },
        "required": [
          "open_count",
          "in_progress_count",
          "pending_count",
          "reopened_count",
          "overdue_count",
          "stagnant_count",
          "due_this_week_count",
          "closed_this_month_count",
          "total_active_count"
        ],
        "title": "DashboardSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueListItem": {
        "properties": {
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "category_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name"
          },
          "days_open": {
            "title": "Days Open",
            "type": "integer"
          },
          "days_since_last_update": {
            "title": "Days Since Last Update",
            "type": "integer"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          },
          "is_overdue": {
            "title": "Is Overdue",
            "type": "boolean"
          },
          "issue_code": {
            "title": "Issue Code",
            "type": "string"
          },
          "last_update_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Update At"
          },
          "next_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action"
          },
          "pic_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pic Name"
          },
          "pic_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pic User Id"
          },
          "priority": {
            "$ref": "#/components/schemas/IssuePriority"
          },
          "raised_date": {
            "format": "date",
            "title": "Raised Date",
            "type": "string"
          },
          "responsible_party_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Party Id"
          },
          "responsible_party_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Party Name"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "issue_code",
          "title",
          "category_id",
          "category_name",
          "responsible_party_id",
          "responsible_party_name",
          "priority",
          "status",
          "raised_date",
          "pic_name",
          "pic_user_id",
          "due_date",
          "days_open",
          "last_update_at",
          "days_since_last_update",
          "next_action",
          "is_overdue",
          "is_archived"
        ],
        "title": "IssueListItem",
        "type": "object"
      },
...
      "IssuePriority": {
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "IssuePriority",
        "type": "string"
      },
...
      "IssueUpdateResponse": {
        "properties": {
          "action_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Owner"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "due_date_after": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date After"
          },
          "due_date_before": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date Before"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "issue_id": {
            "format": "uuid",
            "title": "Issue Id",
            "type": "string"
          },
          "meeting_occurrence_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Occurrence Id"
          },
          "next_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Action"
          },
          "pic_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pic After"
          },
          "pic_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pic Before"
          },
          "progress_percentage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Percentage"
          },
          "status_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status After"
          },
          "status_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Before"
          },
          "target_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "update_date": {
            "format": "date",
            "title": "Update Date",
            "type": "string"
          },
          "update_note": {
            "title": "Update Note",
            "type": "string"
          },
          "void_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Void Reason"
          },
          "voided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voided At"
          },
          "voided_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voided By"
          }
        },
        "required": [
          "id",
          "issue_id",
          "update_date",
          "meeting_occurrence_id",
          "update_note",
          "decision",
          "next_action",
          "action_owner",
          "target_date",
          "progress_percentage",
          "status_before",
          "status_after",
          "due_date_before",
          "due_date_after",
          "pic_before",
          "pic_after",
          "created_by",
          "created_at",
          "voided_at",
          "voided_by",
          "void_reason"
        ],
        "title": "IssueUpdateResponse",
        "type": "object"
      },
...
      "MonthlyTrendPoint": {
        "properties": {
          "closed": {
            "title": "Closed",
            "type": "integer"
          },
          "month": {
            "title": "Month",
            "type": "string"
          },
          "opened": {
            "title": "Opened",
            "type": "integer"
          }
        },
        "required": [
          "month",
          "opened",
          "closed"
        ],
        "title": "MonthlyTrendPoint",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "project-control@example.com",
      "name": "MrHan (Project Control)"
    },
    "description": "Internal API for the **Project Meeting Issue Tracker** (Project Control team).\n\nThe **Issue** is the primary entity; meetings only add follow-up updates. This is\nthe frozen **v1** contract (base path `/api/v1`) used by the Phase 2C frontend.\n\nAuth: Bearer JWT (HS256). Authorization uses the user's *current database role*,\nnot just the token claim. Timestamps are UTC; date-only fields are `YYYY-MM-DD`\nand displayed in Asia/Jakarta by clients.\n",
    "license": {
      "name": "Proprietary \u2014 internal use only"
    },
    "summary": "Meeting-driven issue register and control for the Project Control team.",
    "title": "Project Meeting Issue Tracker",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dashboard/by-category": {
      "get": {
        "operationId": "dashboard_by_category",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CountByLabel"
                  },
                  "title": "Response Dashboard By Category",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "By Category",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/by-responsible-party": {
      "get": {
        "operationId": "dashboard_by_responsible_party",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CountByLabel"
                  },
                  "title": "Response Dashboard By Responsible Party",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "By Responsible Party",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/due-this-week": {
      "get": {
        "operationId": "dashboard_due_this_week",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssueListItem"
                  },
                  "title": "Response Dashboard Due This Week",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Due This Week",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/opened-vs-closed": {
      "get": {
        "operationId": "dashboard_opened_vs_closed",
        "parameters": [
          {
            "in": "query",
            "name": "months",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 24,
              "minimum": 1,
              "title": "Months",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MonthlyTrendPoint"
                  },
                  "title": "Response Dashboard Opened Vs Closed",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Opened Vs Closed",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/overdue": {
      "get": {
        "operationId": "dashboard_overdue",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssueListItem"
                  },
                  "title": "Response Dashboard Overdue",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Overdue",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/recently-updated": {
      "get": {
        "operationId": "dashboard_recently_updated",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssueListItem"
                  },
                  "title": "Response Dashboard Recently Updated",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recently Updated",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/stagnant": {
      "get": {
        "operationId": "dashboard_stagnant",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssueListItem"
                  },
                  "title": "Response Dashboard Stagnant",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stagnant",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/summary": {
      "get": {
        "operationId": "dashboard_summary",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Summary",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/api/v1/issues/{issue_id}/updates": {
      "get": {
        "operationId": "issue_updates_list",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "pattern": "^(asc|desc)$",
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssueUpdateResponse"
                  },
                  "title": "Response Issue Updates List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Updates",
        "tags": [
          "Issue Updates"
        ]
      },
      "post": {
        "operationId": "issue_updates_create",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueUpdateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Update",
        "tags": [
          "Issue Updates"
        ]
      }
    },
    "/api/v1/issues/{issue_id}/updates/{update_id}": {
      "get": {
        "operationId": "issue_updates_get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "update_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Update Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Update",
        "tags": [
          "Issue Updates"
        ]
      }
    },
    "/api/v1/issues/{issue_id}/updates/{update_id}/void": {
      "post": {
        "operationId": "issue_updates_void",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "update_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Update Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueUpdateVoidRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoidResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Void Update",
        "tags": [
          "Issue Updates"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local (frontend proxy to backend)",
      "url": "http://127.0.0.1:5200"
    },
    {
      "description": "Local backend (direct)",
      "url": "http://localhost:8000"
    }
  ],
...
}